Add experimental SIMD vectorization (WIP)
This commit reworks the internals of the game field, getting rid of the
x/y indices by iterating over the array in a linear fashion (we call it
"raw access"). This should gain us two advantages:
- parallelization should work in a more fine-grained manner, as we now
  only need to divide the field in one single dimension
- looping over the field works in a strictly linear fashion, not having to
  take rows/columns into account

These changes are implemented with the goal of experimenting with
SIMD vectorization, which should, in theory, allow us to process the
field much faster by checking for multiple elements for zero at the same
time. This commit features such an experiment, making use of a SIMD vector
sum. Unfortunately, this doesn't work at the moment, slowing down the
calculations by a factor of 5-10 in its current state.